pdf_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pymupdf==1.24.0
httpx==0.27.0
pytesseract==0.3.10
pillow==10.4.0
